==> 00_Consolidate_data_codes.Rmd <==
---
title: "Consolidate_files_code"
output: html_document
date: "2023-08-16"
---

# Combine data
These are the codes to consolidate all the files into a single file. 

There are four behavioral tasks:
- Typing speed task
- Digit Span task
- Visual Pattern task
- Multiplication production task

```{r}
# Load library
library(tidyverse)

# List all CSV files in the "raw_data" folder
csv_files <- list.files("raw_data", pattern = "*.csv", full.names = TRUE)

# Read and consolidate data using tidyverse
consolidated_data <- map_df(csv_files, ~ read_csv(.x, col_types = cols(session = col_character())))

```

## Typing speed task

```{r}
# Remove all personal information and segregate by task
typing_df <- consolidated_data %>% 
  select(typedNumber, typingRT, answerCorr, participant, group) %>% 
  filter(!typedNumber %in% c(41, 39, 59), !is.na(typedNumber))

# Construct the full file path including the folder's name
output_file_path <- file.path("consolidated_data", "consolidated_typing_data.csv")

# Write the data frame to the specified file path
write_csv(typing_df, output_file_path)
```


## Multiplication production task

```{r}
# Remove all personal information and segregate by task
mult_df <- consolidated_data %>% 
  select(typedAnswer, multiplyRT, multiplyCorr, problem, 
         OperandType, productType, participant, group) %>% 
  filter(!productType %in% c('practiceQ'), !is.na(productType))

# Construct the full file path including the folder's name
output_file_path <- file.path("consolidated_data", "consolidated_mult_data.csv")

# Write the data frame to the specified file path
write_csv(mult_df, output_file_path)
```

## Digit span task (Egner)

```{r}
# Remove all personal information and segregate by task
digitspan_egner_df <- consolidated_data %>% 
  select(digitSpan, digits_audio, EgnerDSrt, EgnerDSscore, participant, group) %>% 
  filter(!is.na(EgnerDSrt))


# Construct the full file path including the folder's name
output_file_path <- file.path("consolidated_data", "consolidated_digit-span-egner_data.csv")

# Write the data frame to the specified file path
write_csv(digitspan_egner_df, output_file_path)
```

## Digit span task (WAIS)

```{r}
# Remove all personal information and segregate by task
digitspan_wais_df <- consolidated_data %>% 
  select(digitSpanWAIS, WAISDSrt, WAISDSresponse, WAISDSscore, participant, group) %>% 
  filter(!is.na(WAISDSrt))

# Construct the full file path including the folder's name
output_file_path <- file.path("consolidated_data", "consolidated_digit-span-wais_data.csv")

# Write the data frame to the specified file path
write_csv(digitspan_wais_df, output_file_path)
```
## Visual Pattern task

```{r}
# Remove all personal information and segregate by task
vpt_df <- consolidated_data %>% 
  select(rows, cols, vptCorrect, noOfBlocks, participant, group) %>% 
  filter(!is.na(vptCorrect))

# Construct the full file path including the folder's name
output_file_path <- file.path("consolidated_data", "consolidated_vpt_data.csv")

# Write the data frame to the specified file path
write_csv(vpt_df, output_file_path)
```


# Demographic data from Qualtrics

```{r}
# first demographics data collection
raw_qualtrics1 <- read_csv("demographics_raw_data/raw/stm_demo_part1.csv") %>% 
  slice(-(1:3)) %>% 
  dplyr::select(name, dob_1, dob_2, dob_3, EndDate,
                email, gender, nationality, handedness) %>% 
  distinct(name, .keep_all = TRUE) %>% 
  filter(!is.na(name) & !is.na(dob_1) & !is.na(dob_2) & !is.na(dob_3))

details <- read_csv("demographics_raw_data/raw/participant_details_pnc.csv") %>% 
  dplyr::select(ParticipantID, Email) %>% 
  rename(participant = ParticipantID,
         email = Email) %>% 
  left_join(raw_qualtrics1, by = "email") %>% 
  filter(!is.na(name) & !is.na(dob_1) & !is.na(dob_2) & !is.na(dob_3)) %>% 
  select(-name, -email)

write_csv(details, "demographics_raw_data/cleaned/stm_demo_part1.csv")

# second demographics data collection

stm_demo2 <- read_csv("demographics_raw_data/raw/stm_demo_part2.csv") %>% 
  slice(-1, -2) %>% 
  dplyr::select(EndDate, participant, dob_1, dob_2, dob_3, 
                gender, nationality, handedness) %>% 
  distinct() %>% 
  filter(!is.na(participant) & participant != "" & participant != -99)

write_csv(stm_demo2, "demographics_raw_data/cleaned/stm_demo_part2.csv")
```


```{r}
library(lubridate)

# List all CSV files in the "cleaned" folder
stm_demo1 <- read_csv("demographics_raw_data/cleaned/stm_demo_part1.csv")
stm_demo2 <- read_csv("demographics_raw_data/cleaned/stm_demo_part2.csv")
  
# Combine the two data frames
combined_stm_demo <- bind_rows(stm_demo1, stm_demo2) %>%
  mutate(
    dob = dmy(paste(dob_2, dob_1, dob_3, sep = "-")),  # Create a dob column with Date format
    EndDate = dmy_hm(EndDate),  # Convert EndDate to Date and Time format
    age = as.period(interval(dob, EndDate)) / years(1)  # Calculate age in years
  ) %>%
  select(-dob_1, -dob_2, -dob_3, -EndDate)  # Remove the original date columns dob_1, dob_2, dob_3

# Write the data frame to the specified file path
write_csv(combined_stm_demo, "consolidated_data/combined_demo_data.csv")
```
