==> 03_Data_exploration.Rmd <==
---
title: "EDA"
output: html_document
date: "2023-09-19"
---

## Demographics data for participants

```{r message=FALSE}

# Demographics
library(tidyverse)
demo_df <- read_csv("consolidated_data/combined_demo_data.csv") %>% 
  mutate(handedness = recode(handedness,
    "Right-handed" = "Right-handed",
    "Right" = "Right-handed",
    "Left-handed" = "Left-handed",
    "Left" = "Left-handed"
  )) %>% 
  dplyr::select(participant, age)

# Task performance (main)
library(pastecs)
# Behavioral Tasks
combined_df <- read_csv("combined_tasks.csv")
descriptive_table <- stat.desc(combined_df, norm = TRUE)
write_csv(descriptive_table, "EDA_output/task_performance_metrics_full.csv")

# Behavioral Tasks
combined_df <- read_csv("combined_tasks.csv") %>% 
  dplyr::select(participant, median_RT, perc_accuracy, 
                verbal_span_mean, mean_blocks,
                medium_mult_RT, perc_accuracy_mult) 

descriptive_table <- stat.desc(combined_df, norm = TRUE)
write_csv(descriptive_table, "EDA_output/task_performance_metrics_selected.csv")
```




```{r}
# Final count now is 
nrow(combined_df)
```

## Information for demographic table.
### Age

```{r}

# Count the total number of participants
total_participants <- nrow(combined_df)

# Count the number of females
female_count <- sum(!is.na(combined_df$gender) & combined_df$gender == "Female")

# Calculate the percentage of females
female_percentage <- (female_count / total_participants) * 100

# Calculate the age statistics
mean_age <- mean(combined_df$age, na.rm = TRUE)
sd_age <- sd(combined_df$age, na.rm = TRUE)


# Print the statement
cat("The final sample consisted of", total_participants, "young Malaysians (",
    female_count, "females, ", sprintf("%.2f%%", female_percentage), 
    ") aged", min_age <- min(combined_df$age, na.rm = TRUE), "to", max(combined_df$age, na.rm = TRUE), 
    "years (M =", mean_age, ", SD =", sd_age, ").\n")

```
