==> 04_DESCRIPTIVE-stats-code.Rmd <==
---
title: "DESCRIPTIVE-stats-code"
output: html_document
date: "2023-08-16"
---

# Import libraries
```{r message=FALSE}
library(tidyverse)
library(janitor)

# Load necessary packages
library(dplyr)
library(psych)

library(apaTables)
library(openxlsx)

```

# Data frames

```{r}

demo_df <- read_csv("consolidated_data/combined_demo_data.csv", col_types = cols(
  participant = col_character(),  # Set 'participant' as character type
  gender = col_character(),
  nationality = col_character(),
  handedness = col_character(),
  age = col_double(),
  dob = col_date(format = "%Y-%m-%d")  
))

```

## All Tasks combined

```{r}

# Combine data frames using left_join
combined_df <- read_csv("combined_tasks.csv")


```


# Correlation

```{r}
## Correlation for two digit span tasks
# Convert to numeric
combined_df$verbal_span_wais <- as.numeric(combined_df$verbal_span_wais)
combined_df$verbal_span_egner <- as.numeric(combined_df$verbal_span_egner)

# Calculate the correlation coefficient and remove missing values
correlation <- cor(combined_df$verbal_span_wais, combined_df$verbal_span_egner, use = "pairwise.complete.obs")

# Calculate the p-value for the correlation
correlation_test <- cor.test(combined_df$verbal_span_wais, combined_df$verbal_span_egner, method = "pearson")

# Extract the p-value from the test result
p_value <- correlation_test$p.value

# Print or view the p-value
print(p_value)

# Print or view the correlation coefficient
print(correlation)


```


# Demographics of participants

```{r}
## WRONG CODES HERE

# Final participants in the digit span task
final_participants <- combined_df %>%
  distinct(participant)

combined_df <- demo_df %>% 
  left_join(typing_df,  by = "participant") %>%
  left_join(digit_df, by = "participant") %>%
  left_join(vpt_df, by = "participant") %>% 
  left_join(mult_df, by = "participant") %>% 
  filter(!is.na(median_RT))

final_participants <- combined_df %>%
  distinct(participant)

summary_df <-final_participants %>% 
  

```


## Combine date frames


```{r}

interference_score <- readr::read_csv("consolidated_data/stim_inteference_score.csv")
# add an interference column to mult_df

mult_df <- readr::read_csv("consolidated_data/cleaned_mult-acc_data.csv") %>% 
  left_join(interference_score, by = "problem") 

# Read the CSV file and assign it to combined_rt_df
missing_task_df <- mult_df %>% 
  filter(multiplyCorr == 1) %>% 
  mutate(log10_rt = log10(multiplyRT)) %>%  # Log10 transformed the RT
  group_by(participant) %>%
  summarize_all(~ sum(is.na(.))) %>% 
  filter_at(vars(-participant), any_vars(. != 0))

combined_rt_df <- mult_df %>% 
  # Filter out incorrect trials
  filter(multiplyCorr == 1) %>% 
  # Apply log10 transformation to reaction times
  mutate(log10_rt = log10(multiplyRT))



```
